Enable ending run and flagging job done in checkpoint
By setting the jobIsDone flag to true, the problem indicates two things;
(1) we want to signal to BoxLib to stop the run at the end of the step, and
(2) we want to create a jobIsDone file in the checkpoint.

The purpose of the checkpoint file is that the job submission system can
quickly check for it and not bother to continue resubmitting the run if
the simulation has flagged that the run is done.

The job can also set signalJobIsNotDone to explicitly create a checkpoint
file called jobIsNotDone if indeed jobIsDone is false. The build system
will look for this first and then know not to bother to check the more
complicated criteria like stop_time and max_step, which involve reading
various files. However this requires positive knowledge that the job is
not done.